Configuration refactor and removed unused classes
