--- hermes-consumers/src/main/java/pl/allegro/tech/hermes/consumers/di/ConsumersBinder.java
@@ -1,175 +1,0 @@
-package pl.allegro.tech.hermes.consumers.di;
-
-import org.eclipse.jetty.client.api.Request;
-import org.glassfish.hk2.api.TypeLiteral;
-import org.glassfish.hk2.utilities.binding.AbstractBinder;
-import pl.allegro.tech.hermes.common.admin.zookeeper.ZookeeperAdminCache;
-import pl.allegro.tech.hermes.common.di.factories.UndeliveredMessageLogFactory;
-import pl.allegro.tech.hermes.common.message.undelivered.UndeliveredMessageLog;
-import pl.allegro.tech.hermes.common.metric.executor.InstrumentedExecutorServiceFactory;
-import pl.allegro.tech.hermes.consumers.consumer.ConsumerAuthorizationHandler;
-import pl.allegro.tech.hermes.consumers.consumer.ConsumerMessageSenderFactory;
-import pl.allegro.tech.hermes.consumers.consumer.converter.AvroToJsonMessageConverter;
-import pl.allegro.tech.hermes.consumers.consumer.converter.DefaultMessageConverterResolver;
-import pl.allegro.tech.hermes.consumers.consumer.converter.MessageConverterResolver;
-import pl.allegro.tech.hermes.consumers.consumer.converter.NoOperationMessageConverter;
-import pl.allegro.tech.hermes.consumers.consumer.interpolation.MessageBodyInterpolator;
-import pl.allegro.tech.hermes.consumers.consumer.interpolation.UriInterpolator;
-import pl.allegro.tech.hermes.consumers.consumer.oauth.OAuthAccessTokens;
-import pl.allegro.tech.hermes.consumers.consumer.oauth.OAuthAccessTokensLoader;
-import pl.allegro.tech.hermes.consumers.consumer.oauth.OAuthConsumerAuthorizationHandler;
-import pl.allegro.tech.hermes.consumers.consumer.oauth.OAuthSubscriptionAccessTokens;
-import pl.allegro.tech.hermes.consumers.consumer.oauth.OAuthSubscriptionHandlerFactory;
-import pl.allegro.tech.hermes.consumers.consumer.oauth.OAuthTokenRequestRateLimiterFactory;
-import pl.allegro.tech.hermes.consumers.consumer.oauth.client.OAuthClient;
-import pl.allegro.tech.hermes.consumers.consumer.oauth.client.OAuthHttpClient;
-import pl.allegro.tech.hermes.consumers.consumer.offset.ConsumerPartitionAssignmentState;
-import pl.allegro.tech.hermes.consumers.consumer.offset.OffsetQueue;
-import pl.allegro.tech.hermes.consumers.consumer.rate.ConsumerRateLimitSupervisor;
-import pl.allegro.tech.hermes.consumers.consumer.rate.calculator.OutputRateCalculatorFactory;
-import pl.allegro.tech.hermes.consumers.consumer.rate.maxrate.MaxRatePathSerializer;
-import pl.allegro.tech.hermes.consumers.consumer.rate.maxrate.MaxRateProviderFactory;
-import pl.allegro.tech.hermes.consumers.consumer.rate.maxrate.MaxRateSupervisor;
-import pl.allegro.tech.hermes.consumers.consumer.receiver.ReceiverFactory;
-import pl.allegro.tech.hermes.consumers.consumer.receiver.kafka.BasicMessageContentReaderFactory;
-import pl.allegro.tech.hermes.consumers.consumer.receiver.kafka.KafkaHeaderExtractor;
-import pl.allegro.tech.hermes.consumers.consumer.receiver.kafka.KafkaMessageReceiverFactory;
-import pl.allegro.tech.hermes.consumers.consumer.receiver.kafka.MessageContentReaderFactory;
-import pl.allegro.tech.hermes.consumers.consumer.sender.HttpMessageBatchSenderFactory;
-import pl.allegro.tech.hermes.consumers.consumer.sender.MessageBatchSenderFactory;
-import pl.allegro.tech.hermes.consumers.consumer.sender.MessageSenderFactory;
-import pl.allegro.tech.hermes.consumers.consumer.sender.ProtocolMessageSenderProvider;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.DefaultHttpMetadataAppender;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.DefaultHttpRequestFactoryProvider;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.DefaultSendingResultHandlers;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.EmptyHttpHeadersProvidersFactory;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.HttpClientsFactory;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.HttpClientsWorkloadReporter;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.HttpHeadersProvidersFactory;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.HttpRequestFactoryProvider;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.JettyHttpMessageSenderProvider;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.SendingResultHandlers;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.SslContextFactoryProvider;
-import pl.allegro.tech.hermes.consumers.consumer.sender.http.auth.HttpAuthorizationProviderFactory;
-import pl.allegro.tech.hermes.consumers.consumer.sender.jms.JmsMetadataAppender;
-import pl.allegro.tech.hermes.consumers.consumer.sender.resolver.EndpointAddressResolver;
-import pl.allegro.tech.hermes.consumers.consumer.sender.resolver.InterpolatingEndpointAddressResolver;
-import pl.allegro.tech.hermes.consumers.consumer.trace.MetadataAppender;
-import pl.allegro.tech.hermes.consumers.health.ConsumerHttpServer;
-import pl.allegro.tech.hermes.consumers.health.ConsumerMonitor;
-import pl.allegro.tech.hermes.consumers.message.undelivered.UndeliveredMessageLogPersister;
-import pl.allegro.tech.hermes.consumers.supervisor.ConsumerFactory;
-import pl.allegro.tech.hermes.consumers.supervisor.ConsumersExecutorService;
-import pl.allegro.tech.hermes.consumers.supervisor.ConsumersSupervisor;
-import pl.allegro.tech.hermes.consumers.supervisor.NonblockingConsumersSupervisor;
-import pl.allegro.tech.hermes.consumers.supervisor.process.Retransmitter;
-import pl.allegro.tech.hermes.domain.filtering.MessageFilterSource;
-import pl.allegro.tech.hermes.domain.filtering.MessageFilters;
-import pl.allegro.tech.hermes.domain.filtering.chain.FilterChainFactory;
-
-import javax.inject.Singleton;
-import javax.jms.Message;
-import java.util.Collections;
-
-@Deprecated
-public class ConsumersBinder extends AbstractBinder { //TODO: remove
-
-    @Override
-    protected void configure() {
-        bindSingleton(ConsumerHttpServer.class);//1
-
-        bind(KafkaMessageReceiverFactory.class).in(Singleton.class).to(ReceiverFactory.class);//1
-        bind(BasicMessageContentReaderFactory.class).in(Singleton.class).to(MessageContentReaderFactory.class);//1
-        bind(MessageBodyInterpolator.class).in(Singleton.class).to(UriInterpolator.class);//1
-        bind(InterpolatingEndpointAddressResolver.class).to(EndpointAddressResolver.class).in(Singleton.class);//1
-//        bind(JmsHornetQMessageSenderProvider.class).to(ProtocolMessageSenderProvider.class)
-//                .in(Singleton.class).named("defaultJmsMessageSenderProvider");//1
-        bind(JettyHttpMessageSenderProvider.class).to(ProtocolMessageSenderProvider.class)
-                .in(Singleton.class).named("defaultHttpMessageSenderProvider");//1
-        bind(EmptyHttpHeadersProvidersFactory.class).to(HttpHeadersProvidersFactory.class).in(Singleton.class);//1
-        bind(DefaultSendingResultHandlers.class).to(SendingResultHandlers.class).in(Singleton.class);//1
-        bind(DefaultHttpRequestFactoryProvider.class).to(HttpRequestFactoryProvider.class).in(Singleton.class);//1
-
-        bind("consumer").named("moduleName").to(String.class);//1
-
-        bind(NonblockingConsumersSupervisor.class).in(Singleton.class).to(ConsumersSupervisor.class);//1
-        bindSingleton(MessageSenderFactory.class);//1
-        bindSingleton(HttpAuthorizationProviderFactory.class);//1
-        bindSingleton(ConsumerFactory.class);//1
-        bindSingleton(ConsumerRateLimitSupervisor.class);//1
-        bindSingleton(ConsumersExecutorService.class);//1
-        bindSingleton(OutputRateCalculatorFactory.class);//1
-        bindSingleton(ZookeeperAdminCache.class);//1
-        bindSingleton(InstrumentedExecutorServiceFactory.class);//1
-        bindSingleton(ConsumerMessageSenderFactory.class);//1
-        bindSingleton(NoOperationMessageConverter.class);//1
-        bindSingleton(AvroToJsonMessageConverter.class);//1
-        bind(DefaultMessageConverterResolver.class).in(Singleton.class).to(MessageConverterResolver.class);//1
-        bindSingleton(OffsetQueue.class);//1
-        bindSingleton(ConsumerPartitionAssignmentState.class);//1
-        bindSingleton(Retransmitter.class);//1
-        bindSingleton(ConsumerMonitor.class);//1
-        bindSingleton(SslContextFactoryProvider.class);//1
-        bindSingleton(KafkaHeaderExtractor.class);//1
-        bind(JmsMetadataAppender.class).in(Singleton.class).to(new TypeLiteral<MetadataAppender<Message>>() {
-        });//1
-        bind(DefaultHttpMetadataAppender.class).in(Singleton.class)
-                .to(new TypeLiteral<MetadataAppender<Request>>() {
-                });//1
-
-//        bindFactory(FutureAsyncTimeoutFactory.class).in(Singleton.class)
-//                .to(new TypeLiteral<FutureAsyncTimeout<MessageSendingResult>>() {
-//                });//1
-        bindSingleton(HttpClientsFactory.class);//1
-
-        //  bindFactory(ConsumerNodesRegistryFactory.class).in(Singleton.class).to(ConsumerNodesRegistry.class);//1
-
-        //      bindFactory(SubscriptionCacheFactory.class).in(Singleton.class).to(SubscriptionsCache.class);//1
-//        bindFactory(SubscriptionIdProviderFactory.class).in(Singleton.class).to(SubscriptionIdProvider.class);//
-        //bindFactory(SubscriptionIdsCacheFactory.class).in(Singleton.class).to(SubscriptionIds.class);//1
-//        bindFactory(ConsumerAssignmentCacheFactory.class).in(Singleton.class).to(ConsumerAssignmentCache.class);//1
-//        bindFactory(ClusterAssignmentCacheFactory.class).in(Singleton.class).to(ClusterAssignmentCache.class);//1
-        bindSingleton(HttpClientsWorkloadReporter.class);//1
-
-        bindFactory(UndeliveredMessageLogFactory.class).in(Singleton.class).to(UndeliveredMessageLog.class);//1
-//        bindFactory(ConsumerAssignmentRegistryFactory.class).in(Singleton.class).to(ConsumerAssignmentRegistry.class);//1
-//        bindFactory(SupervisorControllerFactory.class).in(Singleton.class).to(SupervisorController.class);//1
-//        bindFactory(ConsumersRuntimeMonitorFactory.class).in(Singleton.class).to(ConsumersRuntimeMonitor.class);//1
-
-        //HTTPClient
-//        bindFactory(HttpClientFactory.class).in(Singleton.class).to(HttpClient.class).named("http-1-client");//1
-//        bindFactory(OAuthHttpClientFactory.class).in(Singleton.class).to(HttpClient.class).named("oauth-http-client");//1
-//        bindFactory(Http2ClientFactory.class).in(Singleton.class).to(Http2ClientHolder.class);//1
-
-        //MaxRate
-        bindSingleton(MaxRatePathSerializer.class);//1
-        bindSingleton(MaxRateSupervisor.class);//1
-        bindSingleton(MaxRateProviderFactory.class);//1
-//        bindFactory(MaxRateRegistryFactory.class).in(Singleton.class).to(MaxRateRegistry.class);//1
-
-        bindSingleton(UndeliveredMessageLogPersister.class);//1
-//        bindFactory(ByteBufferMessageBatchFactoryProvider.class).in(Singleton.class).to(MessageBatchFactory.class);//1
-        bind(HttpMessageBatchSenderFactory.class).to(MessageBatchSenderFactory.class).in(Singleton.class);//1
-        bindSingleton(FilterChainFactory.class);//1
-        bind(new MessageFilters(Collections.emptyList(), Collections.emptyList())).to(MessageFilterSource.class); //1
-
-        //OAuth
-        bind(OAuthConsumerAuthorizationHandler.class).in(Singleton.class).to(ConsumerAuthorizationHandler.class);//1
-        bindSingleton(OAuthSubscriptionHandlerFactory.class);//1
-        bindSingleton(OAuthTokenRequestRateLimiterFactory.class);//1
-        bindSingleton(OAuthAccessTokensLoader.class);//1
-        bind(OAuthHttpClient.class).in(Singleton.class).to(OAuthClient.class);//1
-        bind(OAuthSubscriptionAccessTokens.class).in(Singleton.class).to(OAuthAccessTokens.class);//1
-//        bindFactory(OAuthProvidersNotifyingCacheFactory.class).in(Singleton.class)
-//                .to(OAuthProvidersNotifyingCache.class);//1
-    }
-
-    private <T> void bindSingleton(Class<T> clazz) {
-        bind(clazz).in(Singleton.class).to(clazz);
-    }
-
-    //1 - stworzony bean
-    //2 - ogarnięte injections
-    //3 - ogarnięte różnice bind/bindFactory
-    //4 - ogarnięty scope
-}
