--- hermes-consumers/src/main/java/pl/allegro/tech/hermes/consumers/di/config/PrimaryBeanCustomizer.java
@@ -1,12 +1,0 @@
-package pl.allegro.tech.hermes.consumers.di.config;
-
-import org.springframework.beans.factory.config.BeanDefinition;
-import org.springframework.beans.factory.config.BeanDefinitionCustomizer;
-
-public class PrimaryBeanCustomizer implements BeanDefinitionCustomizer { //TODO:remove
-
-    @Override
-    public void customize(BeanDefinition bd) {
-        bd.setPrimary(true);
-    }
-}
